Add files via upload
EARTHQUAKES. ANALOGY. FORECAST.
Prediction of earthquakes is currently the most urgent and difficult task for all of Earth Sciences.
In the past 4 November 2014 in Almaty a seminar on seismology. Scientists buried the hopes of millions in the timely warning of an impending earthquake.
The US Geological Survey believes that such a possibility does not exist.
All the tests that were carried out with a test of the strength of the scientific predictions failed, say officials of the German Centre of the Earth Research (GFZ). 
Against this background, the author considers possible and necessary ivate the self-developed technique, theoretically grounded and repeatedly confirmed in practice.


Methods of forecasting major earthquakes based on their counterparts.
As a result of the rapid development of new views in the Earth sciences in the second half of the twentieth century has been revised most of the previous ideas about the world, that ultimately led to the theory of plate tectonic